requirements.txt
#python -m pip install -r requirements.txt
asyncio
aiohttp
azure.eventhub
azure.identity
python-dotenv
confluent_kafka
azure-identity
